Renamed rest of the project to short names
