+++ include/mpp/detail/mat/mat_base.hpp
@@ -21,7 +21,7 @@
 
 #include <mpp/detail/expr/expr_base.hpp>
 #include <mpp/detail/mat/mat_iter.hpp>
-#include <mpp/detail/utility/public.hpp>
+#include <mpp/detail/util/public.hpp>
 #include <mpp/detail/util/util.hpp>
 #include <mpp/util/traits.hpp>
