+++ include/mpp/mat/matfwd.hpp
@@ -20,7 +20,7 @@
 #pragma once
 
 #include <mpp/detail/expr/expr_base.hpp>
-#include <mpp/detail/utility/public.hpp>
+#include <mpp/detail/util/public.hpp>
 #include <mpp/detail/util/util.hpp>
 #include <mpp/util/cfg.hpp>
