+++ tests/CMakeLists.txt
@@ -37,14 +37,14 @@ function(_create_test TEST_FILENAME)
     target_compile_definitions(${TEST_NAME} PRIVATE -DTEST_DATA_PATH="${CMAKE_CURRENT_SOURCE_DIR}/test_data")
 endfunction()
 
-_create_test("customization")
-_create_test("utilities")
-_create_test("iterator")
-_create_test("algorithms")
-_create_test("initialization")
-_create_test("assignment")
-_create_test("arithmetic")
-_create_test("members")
+_create_test("algos")
+_create_test("ariths")
+_create_test("assign")
+_create_test("customize")
+_create_test("init")
+_create_test("iter")
+_create_test("mem_fns")
+_create_test("utils")
 
 if(${PROJECT_NAME_UPPER}_CODE_COVERAGE)
     include("thirdparty/CodeCoverage.cmake")
