+++ tests/src/algos.cpp
@@ -22,8 +22,8 @@
 #include <mpp/algo.hpp>
 #include <mpp/mat.hpp>
 
-#include "../include/custom_allocator.hpp"
-#include "../include/test_utilities.hpp"
+#include "../include/alloc.hpp"
+#include "../include/utils.hpp"
 
 #include <compare>
 #include <string_view>
