+++ tests/src/mem_fns.cpp
